to download remaining parts

+++ m3u8Downloader.py
@@ -6,25 +6,37 @@ import os
 m3u8 = str(sys.argv[1])
 videoName = str(sys.argv[2])
 
-with open('./videos/m3u8/'+m3u8, 'r') as f:
+with open(m3u8, 'r') as f:
     tslist = [line.rstrip() for line in f if line.rstrip().endswith('.ts')]
-print 'Total '+ str(len(tslist)) +' ts files'
+if len(tslist) > 0:
+    print 'Total '+ str(len(tslist)) +' files'
+else:
+    print 'No ts file found.'
+    exit()
 
 index = 1
 tsNames = []
-for tsUrl in tslist[0:3]:
-    res = requests.get(tsUrl, stream=True)
-    if res.status_code == 200:
-        with open('./videos/tmp/'+videoName[0:-3]+'_'+str(index)+'.ts', 'wb') as f:
-            for chunk in res:
-                f.write(chunk)
-        print videoName[0:-3]+'_'+str(index)+' downloaded\r',
-        tsNames.append(videoName[0:-3]+'_'+str(index)+'.ts')
-        index += 1
+for tsUrl in tslist:
+    videoNameTmp = videoName[0:-3]+'_'+str(index)+videoName[-3:]
+    if not os.path.isfile('./videos/tmp/'+videoNameTmp):
+        res = requests.get(tsUrl, stream=True)
+        if res.status_code == 200:
+            with open('./videos/tmp/'+videoNameTmp, 'wb') as f:
+                for chunk in res:
+                    f.write(chunk)
+            print videoNameTmp+' downloaded\r',
+        else:
+            print '\nConnection error'
+            exit()
+    tsNames.append(videoNameTmp)
+    index += 1
 
-with open('./videos/'+videoName, 'wb') as f:
-    for ts in tsNames:
-        with open('./videos/tmp/'+ts, 'rb') as mergefile:
-            shutil.copyfileobj(mergefile, f)
-        os.remove('./videos/tmp/'+ts)
-    print videoName+' merged'
+if index == len(tslist)+1:
+    with open('./videos/'+videoName, 'wb') as f:
+        for ts in tsNames:
+            with open('./videos/tmp/'+ts, 'rb') as mergefile:
+                shutil.copyfileobj(mergefile, f)
+            os.remove('./videos/tmp/'+ts)
+        print videoName+' merged.'
+else:
+    print 'Merge failed, missing files.'
